[server] Cleanup around messagebus and EE split
